projects/hr-document-generator/requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
python-docx>=1.1.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
